requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.8.2,<3.9.0
psycopg2>=2.8.6,<2.8.7